rename: enum 클래스 common 패키지로 이동
- enum 클래스들의 패키지를 common으로 통합하여 구조 정리
- 공통으로 사용하는 enum 위치 일관성 유지
