+++ src/main/java/com/example/outsourcing/common/aop/OrderStatusLogger.java
@@ -1,0 +1,4 @@
+package com.example.outsourcing.common.aop;
+
+public class OrderStatusLogger {
+}
