+++ src/main/java/com/example/outsourcing/common/enums/PaymentMethod.java
@@ -1,4 +1,4 @@
-package com.example.outsourcing.payment.entity;
+package com.example.outsourcing.common.enums;
 
 public enum PaymentMethod {
     CASH,
